+++ 2d-interval-tree-w-top-k/dev/TwoD_IT_w_TopK.h
@@ -2,6 +2,7 @@
 #define TWOD_IT_W_TOPK_H
 
 #include <list>
+#include <inttypes.h>
 #include <string>
 #include <unordered_map>
 #include <unordered_set>
@@ -13,13 +14,13 @@
 class TwoD_Interval {
 public:
   TwoD_Interval() {};
-  TwoD_Interval(const std::string id, const std::string low, const std::string high, const long long max_timestamp)
+  TwoD_Interval(const std::string id, const std::string low, const std::string high, const uint64_t max_timestamp)
     : _id(id), _low(low), _high(high), _max_timestamp(max_timestamp) {};
   
   std::string GetId() const {return _id;};
   std::string GetLowPoint() const {return _low;};
   std::string GetHighPoint() const {return _high;};
-  long long GetMaxTimeStamp() const {return _max_timestamp;};
+  uint64_t GetMaxTimeStamp() const {return _max_timestamp;};
   
   bool operator == (const TwoD_Interval& otherInterval)
     const {return (_id == otherInterval._id);}
@@ -36,7 +37,7 @@ protected:
   std::string _id;
   std::string _low;
   std::string _high;
-  long long _max_timestamp;
+  uint64_t _max_timestamp;
 };
 
 
@@ -47,20 +48,20 @@ public:
   TwoD_IT_w_TopK(const std::string &filename, const bool &sync_from_file);
   ~TwoD_IT_w_TopK();
 
-  void insertInterval(const std::string &id, const std::string &minKey, const std::string &maxKey, const long long &maxTimestamp);
+  void insertInterval(const std::string &id, const std::string &minKey, const std::string &maxKey, const uint64_t &maxTimestamp);
   
   void deleteInterval(const std::string &id);
   void deleteAllIntervals(const std::string &id_prefix);
   
   void getInterval(TwoD_Interval *ret_interval, const std::string &id) const;
-  void topK(std::vector<TwoD_Interval> *ret_value, const std::string &minKey, const std::string &maxKey, const int &k) const;
+  void topK(std::vector<TwoD_Interval> *ret_value, const std::string &minKey, const std::string &maxKey, const uint32_t &k) const;
   
   void sync() const;
 
   void setSyncFile(const std::string &filename);
   void getSyncFile(std::string *filename) const;
-  void setSyncThreshold(const long &threshold);
-  void getSyncThreshold(long *threshold) const;
+  void setSyncThreshold(const uint32_t &threshold);
+  void getSyncThreshold(uint32_t *threshold) const;
   
   void setIdDelimiter(const char &delim);
   void getIdDelimiter(char *delim) const;
@@ -72,8 +73,8 @@ private:
   char id_delim;
   
   std::string sync_file;
-  long sync_threshold;
-  mutable long sync_counter;
+  uint32_t sync_threshold;
+  mutable uint32_t sync_counter;
   
 };
